Move back to the main store
